requirements.txt
colorama==0.4.6
python-dotenv==1.1.0
tabulate==0.9.0
websocket-client==1.8.0
